moving style test to styles folder
