+++ packages/viz/__tests__/styles/SizeCategoriesStyle.spec.ts
@@ -1,8 +1,8 @@
 import { Deck } from '@deck.gl/core';
-import * as mapsResponse from './data-mocks/maps.category.json';
-import { sizeCategoriesStyle } from '../src/lib/style';
-import { CARTOSource } from '../src';
-import { defaultStyles } from '../src/lib/style/default-styles';
+import * as mapsResponse from '../data-mocks/maps.category.json';
+import { sizeCategoriesStyle } from '../../src/lib/style';
+import { CARTOSource } from '../../src';
+import { defaultStyles } from '../../src/lib/style/default-styles';
 
 const FIELD_NAME = 'category';
 const mapStats = mapsResponse.metadata.layers[0].meta.stats;
@@ -19,11 +19,11 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../src', () => ({
+jest.mock('../../src', () => ({
   CARTOSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
 
-jest.mock('../src/lib/style/layer-style', () => ({
+jest.mock('../../src/lib/style/layer-style', () => ({
   pixel2meters: jest.fn().mockImplementation(v => v)
 }));
