+++ packages/viz/__tests__/styles/SizeContinuousStyle.spec.ts
@@ -1,8 +1,8 @@
 import { Deck } from '@deck.gl/core';
-import { sizeContinuousStyle } from '../src/lib/style';
-import { getDefaultSizeRange } from '../src/lib/style/helpers/size-continuous-style';
-import * as mapsResponse from './data-mocks/maps.number.json';
-import { CARTOSource } from '../src';
+import { sizeContinuousStyle } from '../../src/lib/style';
+import { getDefaultSizeRange } from '../../src/lib/style/helpers/size-continuous-style';
+import * as mapsResponse from '../data-mocks/maps.number.json';
+import { CARTOSource } from '../../src';
 
 const FIELD_NAME = 'pct_higher_ed';
 const mapStats = mapsResponse.metadata.layers[0].meta.stats;
@@ -19,11 +19,11 @@ const getMetadata = jest.fn().mockImplementation(() => {
   };
 });
 
-jest.mock('../src', () => ({
+jest.mock('../../src', () => ({
   CARTOSource: jest.fn().mockImplementation(() => ({ getMetadata }))
 }));
 
-jest.mock('../src/lib/style/layer-style', () => ({
+jest.mock('../../src/lib/style/layer-style', () => ({
   pixel2meters: jest.fn().mockImplementation(v => v)
 }));
