+++ packages/viz/__tests__/styles/Style.spec.ts
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/no-unresolved
 import { GeoJsonLayerProps } from '@deck.gl/layers/geojson-layer/geojson-layer';
-import { Style } from '../src/lib/style/Style';
+import { Style } from '../../src/lib/style/Style';
 
 describe('Style', () => {
   describe('Style creation', () => {
